+++ sources/libpassion/include/modules/graphics/ps_graphics_state.h
@@ -48,12 +48,12 @@ PS_PUBLIC_API enum ps_status ps_graphics_is_wireframe(
 
 PS_PUBLIC_API enum ps_status ps_graphics_set_background_color(
         struct ps_passion *this,
-        double red, double green, double blue, double alpha
+        uint8_t red, uint8_t green, uint8_t blue, uint8_t alpha
 );
 
 PS_PUBLIC_API enum ps_status ps_graphics_set_color(
         struct ps_passion *this,
-        double red, double green, double blue, double alpha
+        uint8_t red, uint8_t green, uint8_t blue, uint8_t alpha
 );
 
 PS_PUBLIC_API enum ps_status ps_graphics_set_line_join(
